Local and Federated tests for PCA
