+++ Exareme-Docker/src/mip-algorithms/tests/algorithm_tests/test_calibration_belt.py
