+++ Exareme-Docker/src/mip-algorithms/tests/algorithm_tests/test_pca.py
@@ -1,35 +1,37 @@
 import pytest
-import json
-import requests
-import math
+from pathlib import Path
 import numpy as np
 
-from .lib import vmUrl
+from mipframework.testutils import get_test_params, get_algorithm_result
+from PCA import PCA
 
-endpointUrl = vmUrl + "PCA"
-
-
-def get_test_params():
-    with open("expected/pca_expected.json") as json_file:
-        params = json.load(json_file)["test_cases"]
-    params = [(p["input"], p["output"]) for p in params]
-    return params
+expected_file = Path(__file__).parent / "expected" / "pca_expected.json"
 
 
-@pytest.mark.parametrize("test_input, expected", get_test_params())
-def test_eval(test_input, expected):
-    headers = {"Content-type": "application/json", "Accept": "text/plain"}
-    res = requests.post(endpointUrl, data=json.dumps(test_input), headers=headers)
-    res = json.loads(res.text)
-    res = res["result"][0]["data"]
-    expected = expected[0]
+@pytest.mark.parametrize(
+    "test_input, expected", get_test_params(expected_file, slice(80))
+)
+def test_pca_algorithm_local(test_input, expected):
+    result = get_algorithm_result(PCA, test_input, num_workers=1)
 
-    assert math.isclose(res["n_obs"], expected["n_obs"], rel_tol=1e-5)
-    assert np.isclose(res["eigenvalues"], expected["eigen_vals"], rtol=1e-5).all()
-    for u, v in zip(res["eigenvectors"], expected["eigen_vecs"]):
+    assert int(result["n_obs"]) == int(expected["n_obs"])
+    assert np.isclose(result["eigenvalues"], expected["eigen_vals"], atol=1e-3).all()
+    for u, v in zip(result["eigenvectors"], expected["eigen_vecs"]):
+        assert are_collinear(u, v)
+
+
+@pytest.mark.parametrize(
+    "test_input, expected", get_test_params(expected_file, slice(80, 90))
+)
+def test_pca_algorithm_federated(test_input, expected):
+    result = get_algorithm_result(PCA, test_input, num_workers=10)
+
+    assert int(result["n_obs"]) == int(expected["n_obs"])
+    assert np.isclose(result["eigenvalues"], expected["eigen_vals"], atol=1e-3).all()
+    for u, v in zip(result["eigenvectors"], expected["eigen_vecs"]):
         assert are_collinear(u, v)
 
 
 def are_collinear(u, v):
     cosine_similarity = np.dot(v, u) / (np.sqrt(np.dot(v, v)) * np.sqrt(np.dot(u, u)))
-    return math.isclose(abs(cosine_similarity), 1, rel_tol=1e-5)
+    return np.isclose(abs(cosine_similarity), 1, rtol=1e-5)
