+++ Exareme-Docker/src/mip-algorithms/tests/algorithm_tests/test_pca_exareme.py
@@ -1,0 +1,35 @@
+import pytest
+import json
+import requests
+import math
+import numpy as np
+
+from .lib import vmUrl
+
+endpointUrl = vmUrl + "PCA"
+
+
+def get_test_params():
+    with open("expected/pca_expected.json") as json_file:
+        params = json.load(json_file)["test_cases"]
+    params = [(p["input"], p["output"]) for p in params]
+    return params
+
+
+@pytest.mark.parametrize("test_input, expected", get_test_params())
+def test_eval(test_input, expected):
+    headers = {"Content-type": "application/json", "Accept": "text/plain"}
+    res = requests.post(endpointUrl, data=json.dumps(test_input), headers=headers)
+    res = json.loads(res.text)
+    res = res["result"][0]["data"]
+    expected = expected[0]
+
+    assert math.isclose(res["n_obs"], expected["n_obs"], rel_tol=1e-5)
+    assert np.isclose(res["eigenvalues"], expected["eigen_vals"], rtol=1e-5).all()
+    for u, v in zip(res["eigenvectors"], expected["eigen_vecs"]):
+        assert are_collinear(u, v)
+
+
+def are_collinear(u, v):
+    cosine_similarity = np.dot(v, u) / (np.sqrt(np.dot(v, v)) * np.sqrt(np.dot(u, u)))
+    return math.isclose(abs(cosine_similarity), 1, rel_tol=1e-5)
